feat(getopt): enable mods for source2 engine
The reason for this is to more closely follow the format the earlier metamod/sourcemod workflows use. Metamod is not exclusive to cs2 and may support other source2 titles in the future. Leave in the cs2 flag since there are still cs2-specific mods we will add in later on (cs sharp etc).